13-01-2025 #1
Complete the logic of shared preferences for user location, user can choose a location and the application will save the location in the shared preferences.

UI is also beautified with a location button that directly adjusts the location text to the location selected by the user.

Lastly, unite the homebar with customappbar, the homebar will no longer exist.